apps/particular_task/serializers.py
from rest_framework import serializers

from apps.particular_task.models import Task


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = "__all__"
