requirements.txt
asgiref==3.7.2
assertpy==1.1
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
charset-normalizer==3.2.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.7
Django==4.1.6
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
drf-yasg==1.21.7
dulwich==0.21.5
environ==1.0
exceptiongroup==1.1.3
filelock==3.12.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
more-itertools==10.1.0
msgpack==1.0.5
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2==2.9.7
ptyprocess==0.7.0
pyproject_hooks==1.0.0
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.3
webencodings==0.5.1
zipp==3.16.2
